day 2

+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: aghlimi <aghlimi@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/22 10:18:22 by aghlimi           #+#    #+#             */
-/*   Updated: 2024/10/22 11:18:57 by aghlimi          ###   ########.fr       */
+/*   Updated: 2024/10/22 20:51:45 by aghlimi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 	i = 0;
 	if (!(dst && src && dstsize))
 		return (len);
-	while (src[i] && i < dstsize - 1)
+	while (src[i] && (size_t)i < dstsize - 1)
 	{
 		dst[i] = src[i];
 		i++;
